web/fix: notes footer popover instead of header
