+++ packages/web/app/_components/project/quick-links.tsx
@@ -1,6 +1,7 @@
-import { Box, Card } from '@mantine/core';
-import { useTranslations } from 'next-intl';
 import { FC } from 'react';
+import { useTranslations } from 'next-intl';
+import { Box, Card, Stack } from '@mantine/core';
+import styles from '@/styles/components/quick-links.module.scss';
 
 type Props = {
   projectId: string;
@@ -10,8 +11,10 @@ export const QuickLinks: FC<Props> = (props) => {
   const { t } = useQuickLinks(props);
 
   return (
-    <Box>
-      <Card radius={24} pos="unset"></Card>
+    <Box h="100%">
+      <Card h="100%" radius={24} pos="unset" className={styles.quickLinksCard}>
+        <Stack></Stack>
+      </Card>
     </Box>
   );
 };
