+++ packages/web/styles/components/notes.module.scss
@@ -30,7 +30,6 @@
   position: absolute;
   right: 0;
   top: 0;
-  z-index: 1;
   height: 100%;
   width: 100%;
   min-width: rem(800);
