+++ packages/web/styles/components/quick-links.module.scss
@@ -1,0 +1,3 @@
+.quickLinksCard {
+  border: 1px solid var(--mantine-color-neutral-3);
+}
